+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2020-10-10
+
+## Changed
+- `Transformer` allows returning no message (i.e. to indicate the pipeline stops for a given message)
+- `Piper` deals with no message by transformer by directly responding with NACK & no requeue
+
 ## [0.2.0] - 2020-10-06
 
 ### Added
@@ -29,7 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/tsterker/hopper/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/tsterker/hopper/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/tsterker/hopper/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/tsterker/hopper/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/tsterker/hopper/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/tsterker/hopper/compare/v0.1.0...v0.1.1